+++ yesod-core/ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## 1.6.28.0
 
-* When showing HandlerContent, include a content snippet[#1864](https://github.com/yesodweb/yesod/pull/1864)
+* When showing HandlerContent, include a content snippet [#1864](https://github.com/yesodweb/yesod/pull/1864)
 
 
 ## 1.6.27.0